refactor into more discrete parts
