+++ src/bindings/rust/src/lib.rs
@@ -73,12 +73,15 @@ pub use bindings::{
 };
 
 mod agent;
-mod descriptor;
+mod descriptors;
 mod notify;
+mod utils;
+mod xfer;
 
 pub use agent::*;
-pub use descriptor::*;
+pub use descriptors::*;
 pub use notify::*;
+pub use xfer::*;
 
 /// Errors that can occur when using NIXL
 #[derive(Error, Debug)]
@@ -101,90 +104,6 @@ pub enum NixlError {
     RegDescAddFailed,
 }
 
-/// A safe wrapper around a list of strings from NIXL
-pub struct StringList {
-    inner: NonNull<bindings::nixl_capi_string_list_s>,
-}
-
-impl StringList {
-    /// Returns the number of strings in the list
-    pub fn len(&self) -> Result<usize, NixlError> {
-        let mut size = 0;
-        let status = unsafe { nixl_capi_string_list_size(self.inner.as_ptr(), &mut size) };
-
-        match status {
-            0 => Ok(size),
-            -1 => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns true if the list contains no strings
-    pub fn is_empty(&self) -> Result<bool, NixlError> {
-        Ok(self.len()? == 0)
-    }
-
-    /// Returns the string at the given index
-    pub fn get(&self, index: usize) -> Result<&str, NixlError> {
-        let mut str_ptr = ptr::null();
-        let status = unsafe { nixl_capi_string_list_get(self.inner.as_ptr(), index, &mut str_ptr) };
-
-        match status {
-            0 => {
-                // SAFETY: If status is 0, str_ptr points to a valid null-terminated string
-                let c_str = unsafe { CStr::from_ptr(str_ptr) };
-                Ok(c_str.to_str().unwrap()) // Safe because NIXL strings are valid UTF-8
-            }
-            -1 => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns an iterator over the strings in the list
-    pub fn iter(&self) -> StringListIterator<'_> {
-        StringListIterator {
-            list: self,
-            index: 0,
-            length: self.len().unwrap_or(0),
-        }
-    }
-}
-
-impl Drop for StringList {
-    fn drop(&mut self) {
-        // SAFETY: self.inner is guaranteed to be valid by NonNull
-        unsafe {
-            nixl_capi_destroy_string_list(self.inner.as_ptr());
-        }
-    }
-}
-
-/// An iterator over the strings in a StringList
-pub struct StringListIterator<'a> {
-    list: &'a StringList,
-    index: usize,
-    length: usize,
-}
-
-impl<'a> Iterator for StringListIterator<'a> {
-    type Item = Result<&'a str, NixlError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.index >= self.length {
-            None
-        } else {
-            let result = self.list.get(self.index);
-            self.index += 1;
-            Some(result)
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let remaining = self.length - self.index;
-        (remaining, Some(remaining))
-    }
-}
-
 /// A safe wrapper around NIXL memory list
 pub struct MemList {
     inner: NonNull<bindings::nixl_capi_mem_list_s>,
@@ -195,20 +114,6 @@ impl Drop for MemList {
         // SAFETY: self.inner is guaranteed to be valid by NonNull
         unsafe {
             nixl_capi_destroy_mem_list(self.inner.as_ptr());
-        }
-    }
-}
-
-/// A safe wrapper around NIXL parameters
-pub struct Params {
-    inner: NonNull<bindings::nixl_capi_params_s>,
-}
-
-impl Drop for Params {
-    fn drop(&mut self) {
-        // SAFETY: self.inner is guaranteed to be valid by NonNull
-        unsafe {
-            nixl_capi_destroy_params(self.inner.as_ptr());
         }
     }
 }
@@ -244,7 +149,7 @@ impl RegistrationHandle {
                 let _opt_args = OptArgs::new().unwrap();
                 nixl_capi_deregister_mem(
                     agent.write().unwrap().handle.as_ptr(),
-                    reg_dlist.inner.as_ptr(),
+                    reg_dlist.handle(),
                     _opt_args.inner.as_ptr(),
                 );
             }
@@ -412,100 +317,6 @@ impl Drop for OptArgs {
     }
 }
 
-/// A key-value pair in the parameters
-#[derive(Debug)]
-pub struct ParamPair<'a> {
-    pub key: &'a str,
-    pub value: &'a str,
-}
-
-/// An iterator over parameter key-value pairs
-pub struct ParamIterator<'a> {
-    iter: NonNull<bindings::nixl_capi_param_iter_s>,
-    _phantom: std::marker::PhantomData<&'a ()>,
-}
-
-impl<'a> Iterator for ParamIterator<'a> {
-    type Item = Result<ParamPair<'a>, NixlError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let mut key_ptr = ptr::null();
-        let mut value_ptr = ptr::null();
-        let mut has_next = false;
-
-        // SAFETY: self.iter is guaranteed to be valid by NonNull
-        let status = unsafe {
-            nixl_capi_params_iterator_next(
-                self.iter.as_ptr(),
-                &mut key_ptr,
-                &mut value_ptr,
-                &mut has_next,
-            )
-        };
-
-        match status {
-            0 if !has_next => None,
-            0 => {
-                // SAFETY: If status is 0, both pointers are valid null-terminated strings
-                let result = unsafe {
-                    let key = CStr::from_ptr(key_ptr).to_str().unwrap();
-                    let value = CStr::from_ptr(value_ptr).to_str().unwrap();
-                    Ok(ParamPair { key, value })
-                };
-                Some(result)
-            }
-            -1 => Some(Err(NixlError::InvalidParam)),
-            _ => Some(Err(NixlError::BackendError)),
-        }
-    }
-}
-
-impl Drop for ParamIterator<'_> {
-    fn drop(&mut self) {
-        // SAFETY: self.iter is guaranteed to be valid by NonNull
-        unsafe {
-            nixl_capi_params_destroy_iterator(self.iter.as_ptr());
-        }
-    }
-}
-
-impl Params {
-    /// Returns true if the parameters are empty
-    pub fn is_empty(&self) -> Result<bool, NixlError> {
-        let mut is_empty = false;
-
-        // SAFETY: self.inner is guaranteed to be valid by NonNull
-        let status = unsafe { nixl_capi_params_is_empty(self.inner.as_ptr(), &mut is_empty) };
-
-        match status {
-            0 => Ok(is_empty),
-            -1 => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns an iterator over the parameter key-value pairs
-    pub fn iter(&self) -> Result<ParamIterator<'_>, NixlError> {
-        let mut iter = ptr::null_mut();
-
-        // SAFETY: self.inner is guaranteed to be valid by NonNull
-        let status = unsafe { nixl_capi_params_create_iterator(self.inner.as_ptr(), &mut iter) };
-
-        match status {
-            0 => {
-                // SAFETY: If status is 0, iter was successfully created and is non-null
-                let iter = unsafe { NonNull::new_unchecked(iter) };
-                Ok(ParamIterator {
-                    iter,
-                    _phantom: std::marker::PhantomData,
-                })
-            }
-            -1 => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-}
-
 impl MemList {
     /// Returns true if the memory list is empty
     pub fn is_empty(&self) -> Result<bool, NixlError> {
@@ -582,295 +393,6 @@ impl Iterator for MemListIterator<'_> {
     fn size_hint(&self) -> (usize, Option<usize>) {
         let remaining = self.length - self.index;
         (remaining, Some(remaining))
-    }
-}
-
-/// A safe wrapper around a NIXL transfer descriptor list
-pub struct XferDescList<'a> {
-    inner: NonNull<bindings::nixl_capi_xfer_dlist_s>,
-    _phantom: PhantomData<&'a dyn NixlDescriptor>,
-}
-
-impl<'a> XferDescList<'a> {
-    /// Creates a new transfer descriptor list for the given memory type
-    pub fn new(mem_type: MemType, sorted: bool) -> Result<Self, NixlError> {
-        let mut dlist = ptr::null_mut();
-        let status = unsafe {
-            nixl_capi_create_xfer_dlist(mem_type as nixl_capi_mem_type_t, &mut dlist, sorted)
-        };
-
-        match status {
-            NIXL_CAPI_SUCCESS => {
-                // SAFETY: If status is NIXL_CAPI_SUCCESS, dlist is non-null
-                let inner = unsafe { NonNull::new_unchecked(dlist) };
-                Ok(Self {
-                    inner,
-                    _phantom: PhantomData,
-                })
-            }
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Adds a descriptor to the list
-    pub fn add_desc(&mut self, addr: usize, len: usize, dev_id: u64) -> Result<(), NixlError> {
-        let status = unsafe {
-            nixl_capi_xfer_dlist_add_desc(self.inner.as_ptr(), addr as uintptr_t, len, dev_id)
-        };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(()),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns true if the list is empty
-    pub fn is_empty(&self) -> Result<bool, NixlError> {
-        Ok(self.len()? == 0)
-    }
-
-    /// Returns the number of descriptors in the list
-    pub fn len(&self) -> Result<usize, NixlError> {
-        let mut len = 0;
-        let status = unsafe { nixl_capi_xfer_dlist_len(self.inner.as_ptr(), &mut len) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(len),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns true if any descriptors in the list overlap
-    pub fn has_overlaps(&self) -> Result<bool, NixlError> {
-        let mut has_overlaps = false;
-        let status =
-            unsafe { nixl_capi_xfer_dlist_has_overlaps(self.inner.as_ptr(), &mut has_overlaps) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(has_overlaps),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Clears all descriptors from the list
-    pub fn clear(&mut self) -> Result<(), NixlError> {
-        let status = unsafe { nixl_capi_xfer_dlist_clear(self.inner.as_ptr()) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(()),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Resizes the list to the given size
-    pub fn resize(&mut self, new_size: usize) -> Result<(), NixlError> {
-        let status = unsafe { nixl_capi_xfer_dlist_resize(self.inner.as_ptr(), new_size) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(()),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Add a descriptor from a type implementing NixlDescriptor
-    ///
-    /// # Safety
-    /// The caller must ensure that:
-    /// - The descriptor remains valid for the lifetime of the list
-    /// - The memory region pointed to by the descriptor remains valid
-    pub fn add_storage_desc<D: NixlDescriptor + 'a>(
-        &mut self,
-        desc: &'a D,
-    ) -> Result<(), NixlError> {
-        // Validate memory type matches
-        let desc_mem_type = desc.mem_type();
-        let list_mem_type = unsafe {
-            // Get the memory type from the list by checking first descriptor
-            let mut len = 0;
-            match nixl_capi_xfer_dlist_len(self.inner.as_ptr(), &mut len) {
-                0 => Ok(()),
-                -1 => Err(NixlError::InvalidParam),
-                _ => Err(NixlError::BackendError),
-            }?;
-            if len > 0 {
-                // TODO: Add API to get descriptor memory type
-                MemType::Unknown
-            } else {
-                desc_mem_type
-            }
-        };
-
-        if desc_mem_type != list_mem_type && list_mem_type != MemType::Unknown {
-            return Err(NixlError::InvalidParam);
-        }
-
-        // Get descriptor details
-        let addr = unsafe { desc.as_ptr() } as usize;
-        let len = desc.size();
-        let dev_id = desc.device_id();
-
-        // Add to list
-        self.add_desc(addr, len, dev_id)
-    }
-}
-
-impl Drop for XferDescList<'_> {
-    fn drop(&mut self) {
-        // SAFETY: self.inner is guaranteed to be valid by NonNull
-        unsafe {
-            nixl_capi_destroy_xfer_dlist(self.inner.as_ptr());
-        }
-    }
-}
-
-/// A safe wrapper around a NIXL registration descriptor list
-pub struct RegDescList<'a> {
-    inner: NonNull<bindings::nixl_capi_reg_dlist_s>,
-    _phantom: PhantomData<&'a dyn NixlDescriptor>,
-}
-
-impl<'a> RegDescList<'a> {
-    /// Creates a new registration descriptor list for the given memory type
-    pub fn new(mem_type: MemType, sorted: bool) -> Result<Self, NixlError> {
-        let mut dlist = ptr::null_mut();
-        let status = unsafe {
-            nixl_capi_create_reg_dlist(mem_type as nixl_capi_mem_type_t, &mut dlist, sorted)
-        };
-
-        match status {
-            NIXL_CAPI_SUCCESS => {
-                if dlist.is_null() {
-                    tracing::error!("Failed to create registration descriptor list");
-                    return Err(NixlError::RegDescListCreationFailed);
-                }
-                let ptr = NonNull::new(dlist).ok_or(NixlError::RegDescListCreationFailed)?;
-
-                Ok(Self {
-                    inner: ptr,
-                    _phantom: PhantomData,
-                })
-            }
-            _ => Err(NixlError::RegDescListCreationFailed),
-        }
-    }
-
-    /// Adds a descriptor to the list
-    pub fn add_desc(&mut self, addr: usize, len: usize, dev_id: u64) -> Result<(), NixlError> {
-        let status = unsafe {
-            nixl_capi_reg_dlist_add_desc(self.inner.as_ptr(), addr as uintptr_t, len, dev_id)
-        };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(()),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns true if the list is empty
-    pub fn is_empty(&self) -> Result<bool, NixlError> {
-        Ok(self.len()? == 0)
-    }
-
-    /// Returns the number of descriptors in the list
-    pub fn len(&self) -> Result<usize, NixlError> {
-        let mut len = 0;
-        let status = unsafe { nixl_capi_reg_dlist_len(self.inner.as_ptr(), &mut len) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(len),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Returns true if any descriptors in the list overlap
-    pub fn has_overlaps(&self) -> Result<bool, NixlError> {
-        let mut has_overlaps = false;
-        let status =
-            unsafe { nixl_capi_reg_dlist_has_overlaps(self.inner.as_ptr(), &mut has_overlaps) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(has_overlaps),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Clears all descriptors from the list
-    pub fn clear(&mut self) -> Result<(), NixlError> {
-        let status = unsafe { nixl_capi_reg_dlist_clear(self.inner.as_ptr()) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(()),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Resizes the list to the given size
-    pub fn resize(&mut self, new_size: usize) -> Result<(), NixlError> {
-        let status = unsafe { nixl_capi_reg_dlist_resize(self.inner.as_ptr(), new_size) };
-
-        match status {
-            NIXL_CAPI_SUCCESS => Ok(()),
-            NIXL_CAPI_ERROR_INVALID_PARAM => Err(NixlError::InvalidParam),
-            _ => Err(NixlError::BackendError),
-        }
-    }
-
-    /// Add a descriptor from a type implementing NixlDescriptor
-    ///
-    /// # Safety
-    /// The caller must ensure that:
-    /// - The descriptor remains valid for the lifetime of the list
-    /// - The memory region pointed to by the descriptor remains valid
-    pub fn add_storage_desc(&mut self, desc: &'a dyn NixlDescriptor) -> Result<(), NixlError> {
-        // Validate memory type matches
-        let desc_mem_type = desc.mem_type();
-        let list_mem_type = unsafe {
-            // Get the memory type from the list by checking first descriptor
-            let mut len = 0;
-            match nixl_capi_reg_dlist_len(self.inner.as_ptr(), &mut len) {
-                0 => Ok(()),
-                -1 => Err(NixlError::InvalidParam),
-                _ => Err(NixlError::BackendError),
-            }?;
-            if len > 0 {
-                // TODO: Add API to get descriptor memory type
-                MemType::Unknown
-            } else {
-                desc_mem_type
-            }
-        };
-
-        if desc_mem_type != list_mem_type && list_mem_type != MemType::Unknown {
-            return Err(NixlError::InvalidParam);
-        }
-
-        // Get descriptor details
-        let addr = unsafe { desc.as_ptr() } as usize;
-        let len = desc.size();
-        let dev_id = desc.device_id();
-
-        // Add to list
-        self.add_desc(addr, len, dev_id)
-    }
-}
-
-impl Drop for RegDescList<'_> {
-    fn drop(&mut self) {
-        tracing::trace!("Dropping registration descriptor list");
-        unsafe {
-            nixl_capi_destroy_reg_dlist(self.inner.as_ptr());
-        }
-        tracing::trace!("Registration descriptor list dropped");
     }
 }
 
@@ -952,53 +474,5 @@ impl NixlRegistration for SystemStorage {
         let handle = agent.register_memory(self, opt_args)?;
         self.handle = Some(handle);
         Ok(())
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum XferOp {
-    Read = 0,
-    Write = 1,
-}
-
-/// Methods used for estimating transfer costs
-#[repr(C)]
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum CostMethod {
-    AnalyticalBackend = 0,
-    Unknown = 1,
-}
-
-impl From<u32> for CostMethod {
-    fn from(value: u32) -> Self {
-        match value {
-            0 => CostMethod::AnalyticalBackend,
-            _ => CostMethod::Unknown,
-        }
-    }
-}
-
-/// A handle to a transfer request
-pub struct XferRequest {
-    inner: NonNull<bindings::nixl_capi_xfer_req_s>,
-    agent: Arc<RwLock<AgentInner>>,
-}
-
-// SAFETY: XferRequest can be sent between threads safely
-unsafe impl Send for XferRequest {}
-// SAFETY: XferRequest can be shared between threads safely
-unsafe impl Sync for XferRequest {}
-
-impl Drop for XferRequest {
-    fn drop(&mut self) {
-        unsafe {
-            bindings::nixl_capi_release_xfer_req(
-                self.agent.write().unwrap().handle.as_ptr(),
-                self.inner.as_ptr(),
-            );
-
-            bindings::nixl_capi_destroy_xfer_req(self.inner.as_ptr());
-        }
     }
 }
